Split up a couple of different scripts and namespaced everything
