+++ Assets/Scripts/Battlefield/Camera/CameraTarget.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dfbe935164340f94bb610f2560f55965
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
