+++ Assets/Scripts/Battlefield/CreatureScripts/AAbilitySO.cs
@@ -2,26 +2,28 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public abstract class AAbilitySO : ScriptableObject
-{
-    public enum ActionTypes { Action, BonusAction, Reaction };
-
-    public string AttackName;
-
-    public ActionTypes TypeOfActionRequired;
-
-
-    public abstract void Initialize(GameObject obj);
-    public abstract void TriggerAbility(GameObject target);
-
-    public int Roll(int dice)
+namespace SwordAndBored.Battlefield.CreaturScripts { 
+    public abstract class AAbilitySO : ScriptableObject
     {
-        return Random.Range(1, dice);
-    }
+        public enum ActionTypes { Action, BonusAction, Reaction };
 
-    public int Roll()
-    {
-        return Roll(20);
-    }
+        public string AttackName;
 
+        public ActionTypes TypeOfActionRequired;
+
+
+        public abstract void Initialize(GameObject obj);
+        public abstract void TriggerAbility(GameObject target);
+
+        public int Roll(int dice)
+        {
+            return Random.Range(1, dice);
+        }
+
+        public int Roll()
+        {
+            return Roll(20);
+        }
+
+    }
 }
