+++ Assets/Scripts/Battlefield/CreatureScripts/UnitBase.cs
@@ -3,44 +3,49 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.AI;
 
-public class UnitBase : MonoBehaviour
+
+namespace SwordAndBored.Battlefield.CreaturScripts
 {
-    
 
-    NavMeshAgent agent;
-    [HideInInspector]
-    public Tile currentTile;
-
-    
-    void Awake()
+    public class UnitBase : MonoBehaviour
     {
-        agent = GetComponent<NavMeshAgent>();
-    }
     
 
-    public void MoveTo(Vector3 pos)
-    {
-        if (currentTile)
+        NavMeshAgent agent;
+        [HideInInspector]
+        public Tile currentTile;
+
+    
+        void Awake()
         {
-            currentTile.unitOnTile = null;
-            currentTile = null;
+            agent = GetComponent<NavMeshAgent>();
         }
-        agent.destination = pos;
-    }
+    
 
-    static public int Roll(int dice)
-    {
-        return Random.Range(1, dice);
-    }
+        public void MoveTo(Vector3 pos)
+        {
+            if (currentTile)
+            {
+                currentTile.unitOnTile = null;
+                currentTile = null;
+            }
+            agent.destination = pos;
+        }
 
-    static public int Roll()
-    {
-        return Roll(20);
-    }
+        static public int Roll(int dice)
+        {
+            return Random.Range(1, dice);
+        }
 
-    public void SetTile(Tile tile)
-    {
-        currentTile = tile;
-        tile.unitOnTile = this.gameObject;
+        static public int Roll()
+        {
+            return Roll(20);
+        }
+
+        public void SetTile(Tile tile)
+        {
+            currentTile = tile;
+            tile.unitOnTile = this.gameObject;
+        }
     }
 }
