+++ Assets/Scripts/Battlefield/CreatureScripts/UnitStats.cs
@@ -2,15 +2,19 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class UnitStats : MonoBehaviour
+namespace SwordAndBored.Battlefield.CreaturScripts
 {
+    public class UnitStats : MonoBehaviour
+    {
+        [Header("Player Stats")]
+        public int health;
+        public int attack;
+        public int specialAttack;
+        public int defense;
+        public int specialDefense;
+        public int speedIntit;
+        public int movement;
 
-    public int health;
-    public int attack;
-    public int specialAttack;
-    public int defense;
-    public int specialDefense;
-    public int speedIntit;
-    public int movement;
+    }
 
 }
