+++ Assets/Scripts/Battlefield/GridSystem/GridHolder.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class GridHolder : MonoBehaviour
+{
+    public GameObject[,] tiles = new GameObject[25, 25];
+    public int maxGridSize;
+    
+}
